+++ glue/viewers/common/qt/data_viewer.py
@@ -5,7 +5,7 @@ import os
 from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui
 from glue.core.application_base import ViewerBase
-from glue.core.qt.layer_artist_model import QtLayerArtistContainer, LayerArtistView
+from glue.core.qt.layer_artist_model import QtLayerArtistContainer, LayerArtistWidget
 from glue.external.qt import get_qapp
 from glue.core.qt.mime import LAYERS_MIME_TYPE, LAYER_MIME_TYPE
 from glue.utils.qt import set_cursor
@@ -24,6 +24,8 @@ class DataViewer(ViewerBase, QtGui.QMainWindow):
        * Drag and drop support for adding data
     """
     _layer_artist_container_cls = QtLayerArtistContainer
+    _layer_style_widget_cls = None
+
     LABEL = 'Override this'
 
     def __init__(self, session, parent=None):
@@ -33,8 +35,8 @@ class DataViewer(ViewerBase, QtGui.QMainWindow):
         QtGui.QMainWindow.__init__(self, parent)
         ViewerBase.__init__(self, session)
         self.setWindowIcon(get_qapp().windowIcon())
-        self._view = LayerArtistView()
-        self._view.setModel(self._layer_artist_container.model)
+        self._view = LayerArtistWidget(layer_style_widget_cls=self._layer_style_widget_cls)
+        self._view.layer_list.setModel(self._layer_artist_container.model)
         self._tb_vis = {}  # store whether toolbars are enabled
         self.setAttribute(Qt.WA_DeleteOnClose)
         self.setAcceptDrops(True)
@@ -153,6 +155,10 @@ class DataViewer(ViewerBase, QtGui.QMainWindow):
 
         if self._hub is not None:
             self.unregister(self._hub)
+
+        self._layer_artist_container.clear_callbacks()
+        self._layer_artist_container.clear()
+
         super(DataViewer, self).closeEvent(event)
         event.accept()