+++ CHANGES.md
@@ -112,6 +112,15 @@ v0.7 (unreleased)
 
 * Improved under-the-hood creation of ROIs for Scatter and Histogram Clients. [#676]
 
+* Data viewers can now define a layer artist style editor class that appears
+  under the list of layer artists. [#852]
+
+* Properties of the VisualAttributes class are now callback properties. [#852]
+
+* Add ``glue.utils.qt.widget_properties.connect_value`` function which can take
+  an optional value_range and log option to scale the Qt values to a custom
+  range of values (optionally in log space). [#852]
+
 v0.6 (2015-11-20)
 -----------------